+++ examples/mocha-multi-reporters/mocha-multi-reporters.js
@@ -1,2 +1,0 @@
-// Mocha only looks for reporters in the project directory, so it may miss hoisted packages in npm workspaces
-module.exports = require('mocha-multi-reporters');